+++ packages/codeceptjs-saucelabs/index.js
@@ -1,0 +1,5 @@
+const sauce = require('./lib/codecept.sauce.conf');
+
+ module.exports = {
+     config: { sauce }
+ };